Make compatible with MacOS.

+++ install/src/configure/configure_perl.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-sudo apt-get install -y perl-doc
+sudo apt-get install -y perl-doc >/dev/null 2>&1
 
-sudo cp ../pod2markdown.pl /usr/local/bin/
+if [ -f ../pod2markdown.pl ]; then
+  sudo cp ../pod2markdown.pl /usr/local/bin/
+fi
 
 sudo curl -L https://cpanmin.us 2>/dev/null | perl - App::cpanminus
 sudo cpanm Data::Dump
